Move the wrapper code into common/, and make linalg use @get_xp
